// client/src/types/schema.ts
export interface Attendance {
    id: string;
    userId?: string;
    date: Date;
    checkIn?: Date;
    checkOut?: Date;
    status: string; // present, absent, late, on-break
    location?: string; // lab, icu, desk, resting, lunch
    totalHours?: string;
}